SecureChannel_Server+Client/requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
astroid==1.6.0
attrs==21.4.0
autopep8==2.0.0
backports.functools-lru-cache==1.4
bandit==1.7.4
black==22.10.0
blinker==1.4
boltons==21.0.0
bracex==2.3.post1
Brlapi==0.8.3
certifi==2020.4.5.1
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-option-group==0.5.5
clickclick==20.10.2
colorama==0.4.4
command-not-found==0.3
configparser==3.5.0
connexion==2.14.1
contextlib2==0.6.0.post1
crypto==1.4.1
cryptography==2.9
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.0
distro==1.7.0
distro-info===1.1build1
enum34==1.1.10
face==22.0.0
filelock==3.0.12
Flask==2.2.2
Flask-Pydantic==0.11.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.1
gitdb==4.0.10
GitPython==3.1.29
glom==22.1.0
greenlet==2.0.1
httplib2==0.20.2
idna==2.8
importlib-metadata==1.6.0
importlib-resources==1.4.0
inflection==0.5.1
ipaddress==1.0.23
isort==4.3.4
itermate==1.0.2
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.3.1
logilab-common==1.4.1
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
mccabe==0.6.1
more-itertools==8.10.0
mypy-extensions==0.4.3
Naked==0.1.32
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pathlib2==2.3.5
pathspec==0.10.2
pbr==5.11.0
peewee==3.15.4
pexpect==4.8.0
Pillow==9.0.1
pipenv==2018.11.26
pkcs7==0.1.2
platformdirs==2.5.4
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.1
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==1.10.2
pyflakes==2.2.0
PyGObject==3.42.1
PyJWT==1.7.1
pylint==1.8.3
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.2
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-lsp-jsonrpc==1.0.0
python2-secrets==1.0.5
pytz==2019.3
pyxdg==0.25
PyYAML==6.0
reportlab==3.6.8
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scandir==1.10.0
SecretStorage==2.3.1
semgrep==1.0.0
sha256==0.3
shellescape==3.8.1
simple-crypt==4.1.7
simplejson==3.17.6
singledispatch==3.4.0.3
six==1.14.0
smmap==5.0.0
SQLAlchemy==1.3.10
ssh-import-id==5.11
stevedore==4.1.1
swagger-ui-bundle==0.0.9
systemd-python==234
tinydb==3.15.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing==3.7.4.1
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.6.0
unattended-upgrades==0.1
urllib3==1.25.6
usb-creator==0.3.7
virtualenv==20.0.18
virtualenv-clone==0.5.4
wadllib==1.3.6
wcmatch==8.4.1
Werkzeug==2.2.2
where==1.0.2
wrapt==1.9.0
xdg==5
xkit==0.0.0
zipp==1.2.0
zope.interface==5.4.0
